Ensure preloadedState is immutable on store create
In `getStore` in kea, the `preloadedState` is used as-is, allowing external actors to change the store underneath redux' nose. This specifically WILL happen when using Kea with Next, since the store is created once on the client and new data pushed into it via the Next.js update mechanism. The result is that `react-redux` believes the object has not changed, and components do not update.

Worse, the store is globally available on Window, and the result is that a malicious script would be able to change the store the app believes it is using, without its knowledge.

This change ensures that cannot happen, and retains the redux guarantee of state immutability.